create seeder real

+++ Laravel-9-AdminLTE-master/database/migrations/2023_11_01_113026_create_data_kavling.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('data_kavling', function (Blueprint $table) {
             $table->id();
             $table->string('area_kavling')->unique();
-            $table->integer('harga');
+            $table->bigInteger('harga');
             $table->string('status');
             $table->timestamps();
         });
