+++ Laravel-9-AdminLTE-master/database/migrations/2023_11_23_031514_create_data_transaksi.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('tanggal_check_in');
             $table->date('tanggal_check_out');
             $table->string('area_kavling');
-            $table->float('harga');
+            $table->bigInteger('harga');
             $table->timestamps();
         });
     }
